backend/identification/requirements.txt
torch
transformers
pillow
pandas
requests
numpy
